Input System, Player Movement and Animation
The way that course code handled character movement is simplified too much, so i say a no-no to it.
New Player Input Actions and GameInput scripts - no feeling of playing around with an old input system that Unity had.
New PlayerMovement script is based on slide & collide method that was simplified due to simple nature of player movement in this game (no y-axis movement).
New Animations "Idle" and "Walk" and PlayerAnimator script that are responsive for animating player's model.